Add external gossip healthcheck endpoints to relation data
Add the loadbalancer ip of the consul server to the
relation data with new key external_gossip_healthcheck_endpoints.

consul-clients can use this to update the health check
configuration to check the connectivity to consul-server
over the desired network.

It is expected the loadbalancer ip will be allocated on
the desired network i.e., storage network for storage consul-server.